+++ app/controllers/api/v1/subjects_controller.rb
@@ -14,9 +14,10 @@ class Api::V1::SubjectsController < Api::V1::BaseController
   # def show
   # end
 
-  # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/subjects/1/events.json
-  # def events
-  # end
+  # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/subjects/1/events.json
+  def events
+    @subject.subject_events.each(&:check_coverage)
+  end
 
   # POST /api/v1/projects/1-AUTHENTICATION_TOKEN/subjects.json
   def create