+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 58.0.0
+## 58.0.0 (January 5, 2018)
 
 ### Enhancements
 - **API Changes**